test namespace attrs
