+++ src/tile2net/raster/generate/commandline.py
@@ -1,15 +1,9 @@
-import argparse
 from typing import Optional
 
 import argh
-from toolz import compose_left
-from tile2net.raster.project import Project
-
-# arg = argparse.ArgumentParser().add_argument
-# globals()['arg'] = argh.arg
-import argh
 from argh import arg
 
+from tile2net.raster.util import compose_left
 
 commandline = compose_left(
     argh.expects_obj,
