--- src/tile2net/tileseg/inference/__init__.py
@@ -1,1 +1,0 @@
-from tile2net.tileseg.inference.inference import inference
